openconnect: Manually add dependency to server IP if resolving fails (hack)
Without the added dependency, the connection initially appears to succeed (all through to the 2FA prompt), but later dies with "Write error on DTLS session: Error in the push function." and the RX counter of the interface remaining at zero packets.

There must be a better way to do this (in OpenWrt).